# src/openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: CRMtree API — mobile contract
  version: 1.0.0
  description: |
    The endpoints the CRMtree mobile app (ADR 001) is allowed to depend on.
    The Dart client in `crmtree-mobile` is generated from this file.

    Rules:
    - Old app builds stay installed for weeks. A change that removes or renames
      a field, tightens a type, or adds a required request field is BREAKING
      and needs a new endpoint or version (CI runs `oasdiff` against master).
    - Response objects are open (`additionalProperties: true`): the web API
      returns many more fields than listed. Only the fields documented here
      are part of the contract; the app must ignore the rest.
    - Money values (`value_pln`, `contract_value`) are decimal STRINGS, as
      PostgreSQL NUMERIC serializes — never parse them as binary floats.
    - Every session talks to its own tenant host (`{slug}.crmtree.pl/api`).
      `POST /auth/mobile/login` and `GET /public/app-config` are the only
      calls made on the tenant-less host.
servers:
  - url: https://{tenant}.crmtree.pl/api
    description: Production, per-tenant host
    variables:
      tenant:
        default: app
        description: Tenant slug; `app` for the tenant-less host used before login
  - url: https://int.crmtree.pl/api
    description: Integration environment (tenant comes from the token)
...
security:
  - bearerAuth: []
paths:
...
  /crm/mobile/today:
    get:
      tags: [agenda]
      operationId: getToday
      summary: Everything the "Dziś" screen shows, in one call
      description: |
        Only the signed-in person's own items (plus anyone they currently
        substitute for) — a manager's team view belongs to the calendar. The
        app sends the bounds of its local day, so the result is right in any
        time zone.
      parameters:
        - { name: day_start, in: query, required: true, schema: { type: string, format: date-time }, description: Start of the phone's local day }
        - { name: day_end, in: query, required: true, schema: { type: string, format: date-time }, description: Start of the next local day (exclusive) }
        - { name: month_start, in: query, required: true, schema: { type: string, format: date-time }, description: 'Start of the phone''s local calendar month, for kpis' }
        - { name: month_end, in: query, required: true, schema: { type: string, format: date-time }, description: Start of the next local month (exclusive) }
      responses:
        '200':
          description: Today's agenda, overdue items and leads waiting for a reaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TodayResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /crm/mobile/dashboard:
    get:
      tags: [agenda]
      operationId: getDashboard
      summary: The sales dashboard — figures, the funnel and the sales chart
      description: |
        Counts the leads the user may see: their own for a salesperson, the
        team for a manager, the whole company for an admin. Values are in PLN
        at today's rates. A lead on hold is not in the pipeline; a lead is won
        on the day it was last changed in the "won" stage. The app sends the
        bounds of its local week, month and chart period.
      parameters:
        - { name: week_start, in: query, required: true, schema: { type: string, format: date-time }, description: 'Start of the phone''s local week; the week before it is the comparison' }
        - { name: week_end, in: query, required: true, schema: { type: string, format: date-time }, description: Start of the next local week (exclusive) }
        - { name: month_start, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: month_end, in: query, required: true, schema: { type: string, format: date-time }, description: Start of the next local month (exclusive) }
        - { name: period_end, in: query, required: true, schema: { type: string, format: date-time }, description: 'End of the chart period (exclusive): the start of the phone''s tomorrow' }
        - { name: period_days, in: query, required: true, schema: { type: integer, enum: [7, 30, 90] } }
        - { name: assigned_to, in: query, schema: { type: string, format: uuid }, description: 'One salesperson, for a manager; 403 when outside the user''s scope' }
      responses:
        '200':
          description: The dashboard
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: true
      properties:
        error: { type: string, description: 'Human-readable message (Polish), safe to show' }
        code: { type: string, description: 'Machine-readable code when the app must react, e.g. TOKEN_EXPIRED' }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    TodayResponse:
      type: object
      required: [agenda, overdue, attention, kpis]
      properties:
        kpis: { $ref: '#/components/schemas/TodayKpis' }
        agenda:
          description: Open meetings, calls and tasks dated within the day, earliest first
          type: array
          items: { $ref: '#/components/schemas/TodayItem' }
        overdue:
          description: Open meetings, calls and tasks dated before the day (at most the 50 most recent per source), oldest first
          type: array
          items: { $ref: '#/components/schemas/TodayItem' }
        attention:
          description: The person's leads with unread messages or missed calls, most recently updated first
          type: array
          items: { $ref: '#/components/schemas/AttentionLead' }
    TodayItem:
      type: object
      required: [id, type, title, activity_at, status, source_type, source_id]
      properties:
        id: { type: integer, description: Activity id; unique only together with source_type }
        type: { type: string, enum: [meeting, call, task] }
        title: { type: string }
        body: { type: string, nullable: true }
        activity_at: { type: string, format: date-time }
        duration_min: { type: integer, nullable: true }
        meeting_location: { type: string, nullable: true }
        status: { type: string, enum: [new, open] }
        source_type: { type: string, enum: [lead, partner] }
        source_id: { type: string, description: 'Lead id or partner CRM id, always a string' }
        source_name: { type: string, nullable: true, description: Company name }
    AttentionLead:
      type: object
      required: [lead_id, company, new_email_count, unread_sms_count, unread_whatsapp_count, missed_call_count]
      properties:
        lead_id: { type: string }
        company: { type: string }
        logo_url: { type: string, nullable: true, description: 'Set when the lead has a logo; fetch it from /crm/leads/{id}/logo-img' }
        website: { type: string, nullable: true }
        new_email_count: { type: integer }
        unread_sms_count: { type: integer }
        unread_whatsapp_count: { type: integer }
        missed_call_count: { type: integer }
...
    DashboardResponse:
      type: object
      required: [kpis, funnel, chart]
      properties:
        kpis:
          type: object
          required: [new_leads, new_leads_previous, new_leads_value_pln, new_leads_value_previous_pln, active_leads, pipeline_value_pln, month_won_count, month_won_value_pln]
          properties:
            new_leads: { type: integer, description: Leads created in the week }
            new_leads_previous: { type: integer, description: Leads created in the week before }
            new_leads_value_pln: { type: number }
            new_leads_value_previous_pln: { type: number }
            active_leads: { type: integer, description: 'Leads past "new", not closed and not on hold' }
            pipeline_value_pln: { type: number, description: Value of those leads in whole PLN }
            month_won_count: { type: integer }
            month_won_value_pln: { type: number }
        funnel:
          description: 'Open stages in funnel order, each present even when empty; leads on hold are left out'
          type: array
          items:
            type: object
            required: [stage, count, value_pln]
            properties:
              stage: { type: string, enum: [new, qualification, presentation, offer, negotiation] }
              count: { type: integer }
              value_pln: { type: number }
        chart:
          type: object
          required: [period_days, total_pln, previous_total_pln, points]
          properties:
            period_days: { type: integer }
            total_pln: { type: number, description: Value won in the period }
            previous_total_pln: { type: number, description: Value won in the period of the same length before it }
            points:
              description: 'Value won so far, one number per day of the period, oldest first'
              type: array
              items: { type: number }
...
    TodayKpis:
      type: object
      required: [active_leads, pipeline_value_pln, month_won_value_pln, month_budget_pln]
      properties:
        active_leads: { type: integer, description: 'Own leads past "new", not closed and not on hold' }
        pipeline_value_pln: { type: number, description: 'Value of those leads in PLN, rounded to whole PLN' }
        month_won_value_pln: { type: number, description: 'Value of own leads won in the month, in whole PLN' }
        month_budget_pln: { type: number, description: 'Planned budget for the month in whole PLN; 0 when no budget is set' }